pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metamorph"
version = "0.0.1"
description = "A command line utility for converting files from any format to another"
authors = [{ name = "Francisco Rodrigues", email = "francisco.rodrigues0908@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "filetype~=1.2.0",
    "pillow~=11.1.0",
    "pypandoc_binary~=1.15.0",
    "pdf2image~=1.17.0",
]

[project.scripts]
mm = "mm.cli:main"
